+++ packages/client/screens/add-wallet/add-wallet-states/AddWalletFailureScreen.tsx
@@ -1,49 +1,52 @@
 import React from "react";
-import { View } from "../../components/Themed";
+import { View } from "../../../components/Themed";
 import { StyleSheet } from "react-native";
 import { titleCase } from "@cryptify/common/src/helpers/string_utils";
-import { TitleTextWithIcon } from "../../components/TitleTextWithIcon";
+import { TitleTextWithIcon } from "../../../components/TitleTextWithIcon";
 import { Box, Button, Center } from "native-base";
-import { currenciesDisplayData } from "../../constants/CurrenciesDisplayData";
 import { CurrencyType } from "@cryptify/common/src/domain/currency_type";
-import { AddWalletStatus } from "./add_wallet_status";
-import { faCircleCheckCustom } from "../../components/icons/faCircleCheckCustom";
-import { WalletWithBalance } from "@cryptify/common/src/domain/wallet_with_balance";
-import {CompositeNavigationProp} from "@react-navigation/native";
+import { AddWalletState } from "./add_wallet_state";
+import { CompositeNavigationProp } from "@react-navigation/native";
+import { faCircleXMarkCustom } from "../../../components/icons/faCircleXMarkCustom";
 
 type Props = {
     currencyType: CurrencyType;
-    setStatus: React.Dispatch<React.SetStateAction<AddWalletStatus>>;
+    setState: React.Dispatch<React.SetStateAction<AddWalletState>>;
     walletName: string;
     navigation: CompositeNavigationProp<any, any>;
 };
 
-export default function AddWalletSuccessScreen({ currencyType, setStatus, walletName, navigation }: Props) {
-    const displayData = currenciesDisplayData.find((currency) => currency.type == currencyType)!;
-    const buttonText = `Add another ${titleCase(currencyType)} wallet`;
-
+export default function AddWalletFailureScreen({ currencyType, setState, walletName, navigation }: Props) {
     return (
         <View style={styles.view}>
             <Box paddingTop="130px"></Box>
             <TitleTextWithIcon
-                icon={faCircleCheckCustom}
+                icon={faCircleXMarkCustom}
                 iconSize={96}
                 iconStyles={styles.successIcon}
                 textStyles={styles.title}
                 space={30}
             >
-                Added {titleCase(currencyType)} Wallet
+                Could Not Add {titleCase(currencyType)} Wallet
                 {"\n"}
                 {walletName}
             </TitleTextWithIcon>
             <Center style={{ flex: 1 }}></Center>
-            <Button style={styles.successButton} _text={styles.successButtonText} onPress={() => setStatus(AddWalletStatus.FORM)}>
-                {buttonText}
+            <Button
+                style={styles.successButton}
+                _text={styles.successButtonText}
+                onPress={() => setState(AddWalletState.FORM)}
+            >
+                Try again
             </Button>
-            <Button style={styles.backButton} _text={styles.backButtonText} onPress={() => {
-                navigation.goBack()
-                navigation.goBack()
-            }}>
+            <Button
+                style={styles.backButton}
+                _text={styles.backButtonText}
+                onPress={() => {
+                    navigation.goBack();
+                    navigation.goBack();
+                }}
+            >
                 Back to wallets
             </Button>
             <Box paddingTop="20px"></Box>
@@ -55,7 +58,7 @@ const styles = StyleSheet.create({
     view: {
         flex: 1,
         paddingHorizontal: 15,
-        backgroundColor: "#F0FDF4",
+        backgroundColor: "#FEF2F2",
     },
     title: {
         fontSize: 20,
@@ -63,12 +66,12 @@ const styles = StyleSheet.create({
         textAlign: "center",
     },
     successIcon: {
-        color: "#22C55E",
+        color: "#EF4444",
         fontWeight: "300",
     },
     successButton: {
         marginTop: 7,
-        backgroundColor: "#22C55E",
+        backgroundColor: "#EF4444",
     },
     successButtonText: {
         fontSize: 17,
@@ -79,7 +82,7 @@ const styles = StyleSheet.create({
         backgroundColor: "transparent",
     },
     backButtonText: {
-        color: "#22C55E",
+        color: "#EF4444",
         fontSize: 17,
         lineHeight: 23,
     },
