+++ packages/client/screens/add-wallet/add-wallet-states/add_wallet_state.ts
@@ -1,4 +1,4 @@
-export enum AddWalletStatus {
+export enum AddWalletState {
     FORM,
     LOADING,
     SUCCESS,
